Move App\Probe to App\Slave and rename classes
